airflow_DockerOperator/dist_py/requirements.txt
# dgl==0.6.1
numpy==1.20.3
pandas==1.3.4
PyYAML==6.0
torch==1.11.0
torchmetrics==0.7.3
pyarrow==7.0.0
gunicorn==20.0.4